filter decks in study section
