+++ api/src/resolvers/study.ts
@@ -1,23 +1,14 @@
 import { ApolloError } from 'apollo-server'
-import { endOfToday, isBefore, startOfToday } from 'date-fns'
+import { startOfToday } from 'date-fns'
 import { IResolverObject } from 'graphql-tools'
 
 import { DeckModel, NoteModel, RevisionLogModel } from '../mongo'
-import { FlashCardDocument, FlashCardStatus } from '../mongo/Note'
-import { RevisionLogDocument } from '../mongo/RevisionLog'
 import {
   FlashCardAnswer,
-  MINIMUM_ANSWER_QUALITY,
   answerToQualityValue,
   scheduleFlashCard,
 } from '../utils/scheduler'
-
-const sumByStatus = (logs: RevisionLogDocument[], status: FlashCardStatus) => {
-  return logs.reduce(
-    (total, log) => (log.status === status ? total + 1 : total),
-    0
-  )
-}
+import { studyFlashCardsByDeck } from '../utils/study'
 
 export const queries: IResolverObject = {
   studyFlashCard: async (
@@ -34,31 +25,7 @@ export const queries: IResolverObject = {
       throw new ApolloError('Deck not found')
     }
 
-    const todayLogs = await RevisionLogModel.find({
-      deckId: deck._id,
-      date: { $gte: startOfToday(), $lte: endOfToday() },
-      answerQuality: { $gte: MINIMUM_ANSWER_QUALITY },
-    })
-
-    const numOfNew = sumByStatus(todayLogs, FlashCardStatus.NEW)
-    const numOfLearning = sumByStatus(todayLogs, FlashCardStatus.LEARNING)
-    const numOfReview = sumByStatus(todayLogs, FlashCardStatus.REVIEW)
-
-    const flashCards = (await NoteModel.find({ deckId: deck._id }).exec())
-      .flatMap<FlashCardDocument>((note) => note.cards)
-      .filter(
-        (flashCard) => !flashCard.due || isBefore(flashCard.due, endOfToday())
-      )
-      .filter((flashCard) => {
-        if (flashCard.status === FlashCardStatus.NEW) {
-          return numOfNew < 20
-        } else if (flashCard.status === FlashCardStatus.LEARNING) {
-          return numOfLearning < 100
-        } else if (flashCard.status === FlashCardStatus.REVIEW) {
-          return numOfReview < 50
-        }
-        return false
-      })
+    const flashCards = await studyFlashCardsByDeck(deck._id)
 
     return flashCards[0]
   },
